refactor(auth): split isSmtpConfigured into its own module
Importing the helper from notifications/email.ts pulled nodemailer and
the React Email runtime into every server page that just wanted the env
boolean (login, forgot-password). Move the check to src/lib/smtp.ts so
those pages stay light. Email module re-imports it for its own use.
